Moved IdsConverter to a separate file
